pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "psycopg2-binary>=2.9.10",
]
